backend/requirements/local.txt
-r base.txt

pytest
pytest-django
ruff
